salud_vital/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-filter==24.3
psycopg[binary]==3.2.3
drf-spectacular==0.27.2
python-dotenv==1.0.1
